fixes for fpc 1.1
git-svn-id: http://svn.freepascal.org/svn/lazarus/trunk@3932 4005530d-fff6-0310-9dd1-cebe43e6787f

+++ tools/install/create_fpc_snapshot_tgz.sh
@@ -3,14 +3,27 @@
 #set -x
 set -e
 
-OutputFile=fpcsrc-1.0.7-1.tgz
+OutputFile=$1
+shift
+CVSParams=$@
 
-echo "downloading fpc branch cvs FIXES_1_0_0 ..."
+Usage="$0 outputfilename [cvs params]"
+
+if [ "x$OutputFile" = "x" ]; then
+  echo $Usage
+  exit
+fi
+if [ "x$CVSParams" = "x" ]; then
+  CVSParams="-r FIXES_1_0_0"
+fi
+
+echo downloading cvs $CVSParams ...
 cd /tmp
 rm -rf /tmp/fpc
 export CVSROOT=:pserver:cvs@cvs.freepascal.org:/FPC/CVS
+echo "The password is: cvs"
 cvs login
-cvs -z3 export -r FIXES_1_0_0 fpc
+cvs -z3 export $CVSParams fpc
 
 # pack
 echo "creating tgz ..."
@@ -18,9 +31,6 @@ tar czf fpc_src.tgz fpc
 cd -
 mv /tmp/fpc_src.tgz $OutputFile
 rm -rf /tmp/fpc
-
-echo ""
-echo "NOTE: DON'T FORGET TO PUT THE $OutputFile INTO /usr/src/redhat/SOURCES/"
 
 # end.
